# perf(compose): Remove `await composed()` from hono.ts.