--- a/deno_dist/compose.ts
+++ b/deno_dist/compose.ts
@@ -1,14 +1,22 @@
 import { HonoContext } from './context.ts'
 import type { NotFoundHandler } from './hono.ts'
 
+interface ComposeContext {
+  finalized: boolean
+  res: any
+}
+
 // Based on the code in the MIT licensed `koa-compose` package.
-export const compose = <C>(middleware: Function[], onNotFound?: NotFoundHandler) => {
+export const compose = <C extends ComposeContext>(
+  middleware: Function[],
+  onNotFound?: NotFoundHandler
+) => {
   const middlewareLength = middleware.length
   return (context: C, next?: Function) => {
     let index = -1
     return dispatch(0)
 
-    async function dispatch(i: number): Promise<C> {
+    function dispatch(i: number): C | Promise<C> {
       if (i <= index) {
         throw new Error('next() called multiple times')
       }
@@ -16,20 +24,29 @@ export const compose = <C>(middleware: Function[], onNotFound?: NotFoundHandler)
       index = i
       if (i === middlewareLength && next) handler = next
 
+      let res
+
       if (!handler) {
         if (context instanceof HonoContext && context.finalized === false && onNotFound) {
-          context.res = await onNotFound(context)
+          res = onNotFound(context)
+        }
+      } else {
+        res = handler(context, async () => dispatch(i + 1))
+      }
+
+      if (!(res instanceof Promise)) {
+        if (res && context.finalized === false) {
+          context.res = res
         }
         return context
+      } else {
+        return res.then((res) => {
+          if (res && context.finalized === false) {
+            context.res = res
+          }
+          return context
+        })
       }
-
-      const tmp = handler(context, () => dispatch(i + 1))
-      const res = tmp instanceof Promise ? await tmp : tmp
-
-      if (res && context instanceof HonoContext && context.finalized === false) {
-        context.res = res
-      }
-      return context
     }
   }
 }