@yusukebe 2022-09-02
Hi @usualoma !

Thank you for the PR! I've taken the benchmark for the application just returning "Hello World" on Bun.

![SS](https://user-images.githubusercontent.com/10682/188158194-2d6ab23e-f536-47a3-ae8a-8676a981c6ca.png)

The code with no #493 applied (the code you used in the second benchmark) will be much faster than v2.1.3! This will speed up handlers with some middleware. Nice! I'll merge it.