@yusukebe 2022-09-03
Hi @usualoma !

Sorry to get back on topic. Once again, I benchmarked on Bun. This PR will be effective when multiple middleware is applied to the handler. So, I wrote the following script:


```ts
import { RegExpRouter } from 'hono/router/reg-exp-router'
import { Hono } from '../../src/index'

const app = new Hono({ router: new RegExpRouter() })

for (let i = 0; i < 10; i++) {
  app.use('/with-middleware', async (c, next) => {
    c.set('count', i)
    await next()
    c.get('foo')
  })
}
app.get('/with-middleware', (c) => c.text('GET(/with-middleware)'))
export default app
```

Then, run the scripts on Bun 0.1.11 latest canary build, with each versions:

1. Current `main` branch
2. [fix/logger-called-twice branch](https://github.com/honojs/hono/tree/fix/logger-called-twice/src), not applied this PR yet.
3. v2.1.3

Results are below:

<img width="1040" alt="SS" src="https://user-images.githubusercontent.com/10682/188273072-f8d35205-6ed7-4435-96eb-04e0a9ab9468.png">

<img width="908" alt="SS" src="https://user-images.githubusercontent.com/10682/188273078-7d6843c9-799f-452d-8d89-67dc7017ea3a.png">

Current main branch is slower than the others. Maybe the changes in this PR are having a worse effect.

And, I've notice that all handlers except main handler are `async`, because they are almost middlware. So they have to `await`, and perhaps what this PR is doing is not much different from the previous version. 

What do you think about this? Sorry if I am wrong.