@usualoma 2022-09-03
@yusukebe Thanks!

Your point is spot on.

I did not correctly consider in this PR that "middleware is an `AsyncFunction` and `await` is used".

I think the following PR would solve this problem.
https://github.com/honojs/hono/pull/496